Modified the lensing calculator for NFW SD integration.
Was previously results from Wright and Brainard 2000 where they integrated the profile to infinite radius, but those are now in SD functions with "Full" appended to the name. The new functions only integrate to R_max in terms of R_s, which turns out to be the concentration C.

There is still a huge mismatch in shape and values of GLAMER output values, and these analytical solution. Need to test GLAMER to make sure I am using it right, but the code may not be working.